+++ src/IAFahim.Algebra.Sequence/Prufer.cs
@@ -8,20 +8,20 @@ namespace IAFahim.Algebra.Sequence
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static long Rank(int* seq, int n, int MOD)
         {
-            long r = 0;
-            long pow = 1;
-            for (int i = 0; i < n - 2; i++) pow = pow * n % MOD;
+            long r = 0L;
+            long pow = 1L;
+            for (int i = 0; i < n - 2; i++) pow = (pow * (long)n) % (long)MOD;
             for (int i = 0; i < n - 2; i++)
             {
-                pow = (i == 0) ? Combinatorial.ModPow(n, n - 3, MOD) : pow / n;
-                r = (r + seq[i] * pow) % MOD;
+                pow = (i == 0) ? Combinatorial.ModPow((long)n, (long)(n - 3), (long)MOD) : pow / (long)n;
+                r = (r + (long)seq[i] * pow) % (long)MOD;
             }
-            long result = 0;
-            long pn = 1;
+            long result = 0L;
+            long pn = 1L;
             for (int i = n - 3; i >= 0; i--)
             {
-                result = (result + seq[i] * pn) % MOD;
-                pn = pn * n % MOD;
+                result = (result + (long)seq[i] * pn) % (long)MOD;
+                pn = (pn * (long)n) % (long)MOD;
             }
             return result;
         }
@@ -31,8 +31,8 @@ namespace IAFahim.Algebra.Sequence
         {
             for (int i = n - 3; i >= 0; i--)
             {
-                seq[i] = (int)(rank % n);
-                rank /= n;
+                seq[i] = (int)(rank % (long)n);
+                rank /= (long)n;
             }
         }
     }
